+++ tests/lit.js
@@ -5,17 +5,17 @@ const assert = require('assert');
 describe('lit', () => {
     it('should yield input value', () => {
         const p = pep.lit(5);
-        assert.deepEqual([5], Array.from(p.begin()));
+        assert.deepStrictEqual([5], Array.from(p.begin()));
     });
     
     it('should yield empty strings', () => {
-        assert.deepEqual([''], Array.from(pep.lit('').begin()));
+        assert.deepStrictEqual([''], Array.from(pep.lit('').begin()));
     });
     
     it('should yield falsy values', () => {
-        assert.deepEqual([null], Array.from(pep.lit(null).begin()));
-        assert.deepEqual([false], Array.from(pep.lit(false).begin()));
-        assert.deepEqual([undefined], Array.from(pep.lit().begin()));
+        assert.deepStrictEqual([null], Array.from(pep.lit(null).begin()));
+        assert.deepStrictEqual([false], Array.from(pep.lit(false).begin()));
+        assert.deepStrictEqual([undefined], Array.from(pep.lit().begin()));
     });
     
     it('should yield complex values without transform', () => {