+++ tests/map.js
@@ -6,14 +6,19 @@ describe('map', () => {
 
     it('Should output same value with identity function', () => {
         const p = pep.str(4).map(x =>  x);
-        assert.deepEqual(['4'], Array.from(p.begin()));
+        assert.deepStrictEqual(['4'], Array.from(p.begin()));
+    });
+    
+    it('Should wrap non generators', () => {
+        const p = pep.map('abc', x => x + x);
+        assert.deepStrictEqual(['abcabc'], Array.from(p.begin()));
     });
 
     it('Should map over every element in a sequence', () => {
         const p = pep.map(
             pep.seq(pep.lit(1), pep.lit(2), pep.lit(3)),
             (x) => x * x);
-        assert.deepEqual([1, 4, 9], Array.from(p.begin()));
+        assert.deepStrictEqual([1, 4, 9], Array.from(p.begin()));
     });
     
      it('Should not map over empty', () => {
@@ -22,7 +27,7 @@ describe('map', () => {
         const p = pep.map(
             pep.seq(pep.lit(1), pep.empty, pep.lit(3)),
             (x) => x * x);
-        assert.deepEqual([1, 9], Array.from(p.begin()));
+        assert.deepStrictEqual([1, 9], Array.from(p.begin()));
     });
     
     it('Should map over every element in a sequence without leaking', () => {
@@ -33,6 +38,6 @@ describe('map', () => {
                 (x) => x * x),
             pep.lit(4));
         const n = Array.from(p.begin());
-        assert.deepEqual([5, 1, 4, 9, 4], n);
+        assert.deepStrictEqual([5, 1, 4, 9, 4], n);
     });
 });